+++ CHANGELOG.md
@@ -24,9 +24,10 @@
 
 ## [Unreleased]
 ### Added    
-* Sims can now be run in parallel without dask, using python's in-built `ProcessPoolExecutor`. When `singlethreaded == False`, this is controlled with the `usedask` flag. 
+* Sims can now be run in parallel without dask, using python's in-built `ProcessPoolExecutor`. When `singlethreaded == False`, this is controlled with the `usedask` flag.
+* Can specify `ncores` for running parallel sims. This will override the `n_workers` kwarg in `daskkwargs` if provided.
 ### Changed    
-* When creating an `InVar` with custom values, the nummap for nonnumeric inputs is automatically extracted if not provided.
+* When creating an `InVar` with custom values, the nummap for nonnumeric inputs is automatically extracted if not provided, and no longer raises an error.
 ### Removed    
 
 ## [0.16.3] - 2025-07-16